fix(runtime): close provider lifecycle leaks

+++ process/process_unit_unix.go
@@ -11,14 +11,15 @@ import (
 
 func configureProcessUnit(cmd *exec.Cmd) {
 	cmd.SysProcAttr = &syscall.SysProcAttr{Setpgid: true}
-	cmd.Cancel = func() error {
-		if cmd.Process == nil {
-			return os.ErrProcessDone
-		}
-		err := syscall.Kill(-cmd.Process.Pid, syscall.SIGKILL)
-		if errors.Is(err, syscall.ESRCH) {
-			return os.ErrProcessDone
-		}
-		return err
+}
+
+func cancelProcessUnit(cmd *exec.Cmd) error {
+	if cmd == nil || cmd.Process == nil {
+		return nil
 	}
+	err := syscall.Kill(-cmd.Process.Pid, syscall.SIGKILL)
+	if errors.Is(err, syscall.ESRCH) {
+		return os.ErrProcessDone
+	}
+	return err
 }
